Warning suppression

+++ libs/krypto/encfile.c
@@ -553,7 +553,7 @@ rc_t KEncFileV1BlockEncrypt (KEncFileV1 * self, KEncFileBlock * e)
              ++ bindex)
         {
             /* this goes beyond the end of the data array by design */
-            rindex = (size_t)pw[windex];
+            rindex = pw[windex];
             rindex %= bindex;
 
             self->block.data[bindex] = self->block.data[rindex];
